# fix: keep the logic of the RemoveZeroHex and DecodeHexString methods consistent and add performance testing.