--- a/util/hex.go
+++ b/util/hex.go
@@ -7,10 +7,8 @@ import (
 
 // delete the 0x from the front
 func RemoveZeroHex(s string) []byte {
-	if len(s) > 1 {
-		if s[0:2] == "0x" || s[0:2] == "0X" {
-			s = s[2:]
-		}
+	if strings.HasPrefix(s, "0x") || strings.HasPrefix(s, "0X") {
+		s = s[2:]
 	}
 	if len(s)%2 == 1 {
 		s = "0" + s