`
without0x
BenchmarkDecodeHexString-10    	  176349	      5912 ns/op
with0x
BenchmarkDecodeHexString-10    	  152390	      6741 ns/op
without0x
BenchmarkDecodeHexStringBackupWithout0x-10    	  189400	      5731 ns/op
with0x
BenchmarkDecodeHexStringBackupWith0x-10    	  173818	      6307 ns/op`
`
I have a question. Why write two methods with the same functionality, but one of them doesn't perform error handling?