@yanminmin 2023-07-24
Take a look at the golang kernel code,it is very similar to your code:

// HasPrefix tests whether the string s begins with prefix.
func HasPrefix(s, prefix string) bool {
	return len(s) >= len(prefix) && s[0:len(prefix)] == prefix
}